11_pyav_prober/pyproject.toml
[project]
name = "49-uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "av>=14.3.0",
    "numpy>=2.2.4",
    "python-dotenv>=1.0.1",
    "python-json-logger>=3.2.1",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "flake8>=7.1.2",
    "flake8-2020>=1.8.1",
    "flake8-bugbear>=24.12.12",
    "pytest>=8.3.4",
]
